cleanup

+++ onyxia.sh
@@ -1,9 +1,13 @@
 #!/bin/sh
+
+REPO_URL="https://github.com/RLesur/MachineLearning.git --branch onyxia"
+IPYNB_PATH=matrice_de_confusion_ROC_AUC.ipynb
+
 # Clone the repository in /home/jovyan/work
-git clone https://github.com/RLesur/MachineLearning.git --branch onyxia /home/jovyan/work
+git clone $REPO_URL /home/jovyan/work
 
 # Install dependencies
-pip install -r /home/jovyan/work/requirements.txt
+[ -f /home/jovyan/work/requirements.txt ] && pip install -r /home/jovyan/work/requirements.txt
 
 # Open a given notebook
-echo "\nc.NotebookApp.default_url = '/tree/work/matrice_de_confusion_ROC_AUC.ipynb'" >> /home/jovyan/.jupyter/jupyter_notebook_config.py
+[ -z "$IPYNB_PATH" ] || echo "\nc.NotebookApp.default_url = '/tree/work/${IPYNB_PATH}'" >> /home/jovyan/.jupyter/jupyter_notebook_config.py
